Scheme location choices save successfully
