+++ app/views/application/scheme/wherecouldyouwork.scala.html
@@ -8,7 +8,6 @@
 @import helpers.CSRFieldConstructor._
 @import views.html.widgets.{guard, guardWithLabel, inputCheckboxGroup, radioBox, select_guard}
 
-
 @main_template(title = "Where could you work from?", pageForms = Seq(locationForm)) {
 
     @helper.form(action = routes.SchemeController.submitLocations(), 'novalidate -> "novalidate") {
@@ -33,7 +32,7 @@
                     }
 
                     <section>
-                        <div class="scheme-input-container">
+                        <div id="scheme-input-container" class="scheme-input-container" data-hasALevels="@viewModel.hasALevels" data-hasStemALevels="@viewModel.hasStemALevels">
                             <div id="scrollingList" class="scrolling-panel max-height-25">
                                 <ul class="schemes-list" id="listOfLocations">
                                 </ul>
@@ -60,9 +59,7 @@
             </section>
         </div>
         <div class="form-group">
-            <a href="scheme-selection-new-02.html" class="button" id="locationsSaveBtn">Save and continue</a>
+            <button type="submit" class="button" id="locationsSaveBtn">Save and continue</button>
         </div>
     }
-
-
 }
