+++ app/views/govuk_template.scala.html
@@ -122,12 +122,11 @@
                 </div>
             </footer> <!-- end of footer -->
 
-            <!-- Adapted from: https://github.com/hmrc/play-ui/blob/7a5a10c63e9623ca05346aa2fe3ad4a8457b2598/src/main/twirl/uk/gov/hmrc/play/views/layouts/footer.scala.html -->
             @if(FrontendAppConfig.analyticsToken != "N/A") {
                 <script type="text/javascript">
                     (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
                     (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-                    m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
+                    m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.p7arentNode.insertBefore(a,m)
                     })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
                     ga('create', '@FrontendAppConfig.analyticsToken', '@FrontendAppConfig.analyticsHost');
                     ga('send', 'pageview', { 'anonymizeIp': true });
@@ -138,7 +137,6 @@
             <script src='@routes.Assets.versioned("js/vendor/jquery-ui-1.10.4.custom.min.js")'></script>
             <!-- TODO: Move these two location scripts elsewhere -->
             <script src='@routes.Assets.versioned("js/location-finder.js")'></script>
-            <script src='@routes.Assets.versioned("js/preference-selector.js")'></script>
             <script src='@routes.Assets.versioned("js/vendor/fastclick-1.0.6.min.js")'></script>
             <script src='@routes.Assets.versioned("js/scripts.js")'></script>
