+++ app/views/main_template.scala.html
@@ -1,7 +1,6 @@
-@(title: String, pageForms: Seq[Form[_]] = Nil, notification: Option[(helpers.NotificationType,String)] = None
-)(mainContent: Html)(implicit request: Request[_], flash: Flash, user: Option[CachedData], feedbackUrl: String)
+@(title: String, pageForms: Seq[Form[_]] = Nil, notification: Option[(helpers.NotificationType,String)] = None)(mainContent: Html)(implicit request: Request[_], flash: Flash, user: Option[CachedData], feedbackUrl: String)
 
-@import views.html.template.{errorSummary, flashMessage}
+@import views.html.template.{ errorSummary, flashMessage }
 
 @govuk_template(title = title, user = user) {
     @flashMessage(notification)
